create utils directory
