+++ Asterix/optical_systems/pupil.py
@@ -8,9 +8,9 @@ import skimage.transform
 
 from Asterix import model_dir
 from Asterix.optical_systems import OpticalSystem
-import Asterix.processing_functions as proc
+import Asterix.utils.processing_functions as proc
 import Asterix.optics.phase_amplitude_functions as phase_ampl
-import Asterix.save_and_read as saveread
+from Asterix.utils import save_plane_in_fits
 
 
 class Pupil(OpticalSystem):
@@ -228,7 +228,7 @@ class Pupil(OpticalSystem):
 
         if save_all_planes_to_fits == True:
             name_plane = 'EF_PP_before_pupil' + '_wl{}'.format(int(wavelength * 1e9))
-            saveread.save_plane_in_fits(dir_save_all_planes, name_plane, entrance_EF)
+            save_plane_in_fits(dir_save_all_planes, name_plane, entrance_EF)
 
         if len(self.pup.shape) == 2:
             exit_EF = entrance_EF * self.pup
@@ -244,6 +244,6 @@ class Pupil(OpticalSystem):
 
         if save_all_planes_to_fits == True:
             name_plane = 'EF_PP_after_pupil' + '_wl{}'.format(int(wavelength * 1e9))
-            saveread.save_plane_in_fits(dir_save_all_planes, name_plane, exit_EF)
+            save_plane_in_fits(dir_save_all_planes, name_plane, exit_EF)
 
         return exit_EF
