+++ Asterix/wfsc/thd_quick_invert.py
@@ -7,7 +7,7 @@ import time
 import numpy as np
 
 from astropy.io import fits
-from .wf_control_functions import invert_svd
+from Asterix.utils import invert_svd
 
 def THD_quick_invert(Nbmodes, name_active_DM, matrix_directory, regularization):
     """ --------------------------------------------------
